tools: update inspector_protocol to 97d3146
To apply the diff cleanly, ffb34b6 had to be reverted.

Refs: https://chromium.googlesource.com/deps/inspector_protocol/+log

PR-URL: https://github.com/nodejs/node/pull/39694

+++ tools/inspector_protocol/encoding/encoding.h
@@ -427,7 +427,7 @@ Status AppendString8EntryToCBORMap(span<uint8_t> string8_key,
                                    std::string* cbor);
 
 namespace internals {  // Exposed only for writing tests.
-int8_t ReadTokenStart(span<uint8_t> bytes,
+size_t ReadTokenStart(span<uint8_t> bytes,
                       cbor::MajorType* type,
                       uint64_t* value);
